Move shared functions to their own module (from __main__).
